requirements.txt
asgiref==3.5.2
Django==4.1.2
django-cors-headers==3.3.0
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
gunicorn==20.1.0
importlib-metadata==5.0.0
Markdown==3.4.1
psycopg2==2.9.3
pydotplus==2.0.2
pyparsing==3.0.9
pytz==2020.1
sqlparse==0.3.1
tzdata==2022.4
xlrd==2.0.1
zipp==3.8.1
